move util to auth folder
